refactor: Phase 1 — scaffold findata/ package from corporate_db
Lift corporate_db into the warehouse package:
- findata/db/base.py        — the single DeclarativeBase
- findata/db/session.py     — get_engine / get_session / init_db
- findata/db/migrations/    — the one Alembic tree; alembic.ini moved to repo root
- findata/models/           — exchange, company, insider ORM models
- findata/config.py         — DATABASE_URL / ECHO_SQL (sqlite fallback now resonance.db)
Remove corporate_db/; repoint all importers (descriptions/populate_db.py, Makefile,
`python -m findata`). Move the old DDL reference to docs/schema.sql. Update README,
CLAUDE.md, and the cleanup plan.
