+++ alembic.ini
@@ -1,24 +1,20 @@
-# Alembic configuration for corporate_db
+# Alembic configuration for findata
 # =========================================
-# This file is used by the `alembic` command-line tool.
-# Run migrations from the `corporate_db/` directory:
+# This file is used by the `alembic` command-line tool.  Run from the repo root:
 #
-#   cd corporate_db
-#   alembic upgrade head
+#   alembic upgrade head            # apply all pending migrations
+#   alembic downgrade -1            # roll back the last migration
+#   alembic revision --autogenerate -m "message"
 #
-# Or from the repo root:
-#
-#   alembic -c corporate_db/alembic.ini upgrade head
-#
-# The database URL is normally set via the DATABASE_URL environment variable.
-# The placeholder below is overridden by env.py at runtime.
+# The database URL is normally set via the DATABASE_URL environment variable;
+# the value below is only a fallback and is overridden by env.py at runtime.
 
 [alembic]
 # Path to the migration scripts, relative to this file's location
-script_location = db/migrations
+script_location = findata/db/migrations
 
 # SQLAlchemy URL — overridden at runtime by env.py reading DATABASE_URL env var
-sqlalchemy.url = sqlite:///corporate_db.sqlite3
+sqlalchemy.url = sqlite:///resonance.db
 
 # Template for new migration files
 # file_template = %%(rev)s_%%(slug)s
