--- corporate_db/__init__.py
@@ -1,26 +1,0 @@
-"""
-corporate_db — top-level package.
-
-This module makes the most commonly used symbols available at the package
-level so that callers can do::
-
-    from corporate_db import Company, Exchange, get_session
-
-For more granular imports, use the sub-packages directly::
-
-    from corporate_db.models.company import Company
-    from corporate_db.db.connection import get_session, init_db
-"""
-
-from corporate_db.models import Base, Company, Exchange, Insider
-from corporate_db.db.connection import get_engine, get_session, init_db
-
-__all__ = [
-    "Base",
-    "Company",
-    "Exchange",
-    "Insider",
-    "get_engine",
-    "get_session",
-    "init_db",
-]
