--- corporate_db/__main__.py
@@ -1,12 +1,0 @@
-"""Executable entry point for the ``corporate_db`` package.
-
-Run with::
-
-    python -m corporate_db
-"""
-
-from corporate_db.db.connection import main
-
-
-if __name__ == "__main__":
-    main()
