+++ docs/CLEANUP_PLAN.md
@@ -1,6 +1,12 @@
 # Repo Cleanup Plan — Financial Data Warehouse
 
-Status: **proposed** (decisions captured below; no code changed yet)
+Status: **in progress** — Phase 0 and Phase 1 done. Decisions captured below.
+
+Progress:
+- ✅ Phase 0 — hygiene: notebooks consolidated under `notebooks/`, docs refreshed, `.env.example` added, merged branches pruned.
+- ✅ Phase 1 — `findata/` package created: `corporate_db/` moved in (`findata/db/{base,session}.py`, `findata/models/`, `findata/config.py`), single Alembic tree at `findata/db/migrations/` with `alembic.ini` at the repo root, `corporate_db/` removed, all importers updated (`descriptions/populate_db.py`, `Makefile`).
+- ⬜ Phase 2 — port news tables to ORM.
+- ⬜ Phases 3–7 — see below.
 
 ## Goal
