+++ findata/__init__.py
@@ -1,0 +1,25 @@
+"""
+findata — the financial data warehouse package.
+
+Re-exports the most commonly used symbols::
+
+    from findata import Company, Exchange, get_session, init_db
+
+For granular imports, use the sub-packages directly::
+
+    from findata.models.company import Company
+    from findata.db.session import get_session, init_db
+"""
+
+from findata.models import Base, Company, Exchange, Insider
+from findata.db.session import get_engine, get_session, init_db
+
+__all__ = [
+    "Base",
+    "Company",
+    "Exchange",
+    "Insider",
+    "get_engine",
+    "get_session",
+    "init_db",
+]
