+++ findata/db/__init__.py
@@ -1,11 +1,11 @@
 """
-corporate_db.db package.
+findata.db package.
 
 Re-exports the most commonly used database helpers so that callers can do::
 
-    from corporate_db.db import get_session, init_db
+    from findata.db import get_session, init_db
 """
 
-from .connection import get_engine, get_session, init_db
+from .session import get_engine, get_session, init_db
 
 __all__ = ["get_engine", "get_session", "init_db"]
