+++ findata/models/__init__.py
@@ -1,11 +1,11 @@
 """
-corporate_db.models package.
+findata.models package.
 
 Importing this package makes all ORM models available and ensures that
 Alembic's autogenerate feature can discover them via ``Base.metadata``.
 """
 
-from .base import Base
+from findata.db.base import Base
 from .exchange import Exchange
 from .company import Company
 from .insider import Insider
