+++ findata/requirements.txt
@@ -1,6 +1,6 @@
-# Dependencies for the corporate_db module
+# Dependencies for the findata package
 # ==========================================
-# Install with:  pip install -r corporate_db/requirements.txt
+# Install with:  pip install -r findata/requirements.txt
 
 sqlalchemy>=2.0
 alembic>=1.13
